@@ -8,12 +8,7 @@
     "jquery": "^3.5.1",
     "merge": "^1.2.1",
     "oidc-client": "^1.9.0",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.4.4",
-    "reactstrap": "^8.4.1",
+    "react": "^16.8.0",
     "rimraf": "^2.6.2"
   },
   "devDependencies": {
