@@ -1,14 +1,25 @@
-import React, { Component } from 'react';
+import React, {useState} from 'react';
+import ItemList from "./ItemComponent/ItemList";
 
-export class Home extends Component {
-  static displayName = Home.name;
+export function Home() {
+    const [items, setItems] = useState([
+    { id: 1, type: "fruit", title: "Апельсин" },
+    { id: 2, type: "fruit", title: "Apple" },
+    { id: 3, type: "fruit", title: "Розмарин" },
+])
 
-  
-  
-  render () {
-    return (
+      function toggleItem(id) {
+    setItems(items.map(item => {
+        if (item.id === id) {
+            item.title = "Потрачено";
+        }
+        return item
+    }))
+      }
+      
+      return (
         <div className="board"> 
+          <ItemList items={items}  onToggle={toggleItem}/>
         </div>
     );
   }
-}
